refactor(snmp_topology): split v1 topology renderer
