+++ src/go/plugin/go.d/collector/snmp_topology/func_topology_v1_presentation.go
@@ -1,0 +1,435 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+package snmptopology
+
+import topologyv1 "github.com/netdata/netdata/go/plugins/pkg/topology/v1"
+
+func snmpTopologyV1ActorTypes() map[string]topologyv1.ActorType {
+	types := make(map[string]topologyv1.ActorType)
+	addDevice := func(id, label, colorSlot, icon string) {
+		types[id] = topologyv1.ActorType{
+			Layer:             "network",
+			Identity:          []string{"id"},
+			MergeIdentity:     []string{"chassis_ids", "mac_addresses", "ip_addresses", "sys_name"},
+			AggregationScopes: []string{"device", "network"},
+			Search: &topologyv1.ActorSearchPolicy{
+				Columns: []string{"display_name", "sys_name", "management_ip", "vendor", "model"},
+			},
+			Presentation: &topologyv1.ActorPresentation{
+				Label:     label,
+				Role:      "actor",
+				Icon:      icon,
+				ColorSlot: colorSlot,
+				Border:    &topologyv1.BorderPresentation{Enabled: new(true)},
+				Size:      &topologyv1.ActorSizePresentation{Mode: "link_count", Scale: "emphasized"},
+				Layout:    &topologyv1.ActorLayoutPresentation{Repulsion: "stronger"},
+				LabelPolicy: &topologyv1.LabelPolicy{
+					Columns:   []string{"display_name", "sys_name"},
+					Fallback:  "type_label",
+					MaxLength: 80,
+					Array:     "reject",
+				},
+				Ports: &topologyv1.ActorPortsPresentation{
+					ShowBullets: true,
+					Sources: []topologyv1.PortSourcePresentation{
+						{
+							Source:       "actor_table",
+							Table:        "actor_ports",
+							ActorColumn:  "actor",
+							NameColumn:   "name",
+							TypeColumn:   "topology_role",
+							DefaultType:  "topology",
+							StatusColumn: "oper_status",
+							ModeColumn:   "link_mode",
+							RoleColumn:   "topology_role",
+						},
+					},
+				},
+				Modal: snmpTopologyV1DeviceModal(),
+			},
+		}
+	}
+
+	addDevice("device", "Device", "primary", "server")
+	addDevice("router", "Router", "primary", "router")
+	addDevice("switch", "Switch", "primary", "switch")
+	addDevice("firewall", "Firewall", "warning", "firewall")
+	addDevice("access_point", "Access Point", "info", "access_point")
+	addDevice("server", "Server", "secondary", "server")
+	addDevice("storage", "Storage", "secondary", "storage")
+	addDevice("load_balancer", "Load Balancer", "info", "load_balancer")
+	addDevice("printer", "Printer", "neutral", "printer")
+	addDevice("phone", "Phone", "neutral", "phone")
+	addDevice("ups", "UPS", "structural", "ups")
+	addDevice("camera", "Camera", "neutral", "camera")
+
+	types[snmpTopologyV1ActorEndpoint] = topologyv1.ActorType{
+		Layer:             "network",
+		Identity:          []string{"id"},
+		MergeIdentity:     []string{"mac_addresses", "ip_addresses"},
+		AggregationScopes: []string{"endpoint", "network"},
+		Search:            &topologyv1.ActorSearchPolicy{Columns: []string{"display_name"}},
+		Presentation: &topologyv1.ActorPresentation{
+			Label:     "Inferred endpoint",
+			Role:      "endpoint",
+			Icon:      "remote-endpoint",
+			ColorSlot: "derived",
+			Border:    &topologyv1.BorderPresentation{Enabled: new(true)},
+			Size:      &topologyv1.ActorSizePresentation{Mode: "fixed", Scale: "compact"},
+			Layout:    &topologyv1.ActorLayoutPresentation{Repulsion: "weaker"},
+			LabelPolicy: &topologyv1.LabelPolicy{
+				Columns:   []string{"display_name"},
+				Fallback:  "type_label",
+				MaxLength: 80,
+				Array:     "reject",
+			},
+			Modal: snmpTopologyV1EndpointModal(),
+		},
+	}
+	types[snmpTopologyV1ActorSegment] = topologyv1.ActorType{
+		Layer:             "network",
+		Identity:          []string{"id"},
+		MergeIdentity:     []string{"id"},
+		ParentIdentity:    []string{"parent_devices"},
+		AggregationScopes: []string{"segment", "network"},
+		Search:            &topologyv1.ActorSearchPolicy{Enabled: new(false)},
+		Presentation: &topologyv1.ActorPresentation{
+			Label:     "Network segment",
+			Role:      "group",
+			Icon:      "segment",
+			ColorSlot: "dim",
+			Size:      &topologyv1.ActorSizePresentation{Mode: "fixed", Scale: "compact"},
+			Layout:    &topologyv1.ActorLayoutPresentation{Repulsion: "weakest"},
+			LabelPolicy: &topologyv1.LabelPolicy{
+				Columns:   []string{"display_name"},
+				Fallback:  "type_label",
+				MaxLength: 80,
+				Array:     "reject",
+			},
+			Modal: snmpTopologyV1EndpointModal(),
+		},
+	}
+	types["custom"] = topologyv1.ActorType{
+		Layer:             "custom",
+		Identity:          []string{"id"},
+		MergeIdentity:     []string{"id"},
+		AggregationScopes: []string{"network"},
+		Search:            &topologyv1.ActorSearchPolicy{Columns: []string{"display_name"}},
+		Presentation: &topologyv1.ActorPresentation{
+			Label:     "Custom",
+			Role:      "actor",
+			Icon:      "service",
+			ColorSlot: "neutral",
+			LabelPolicy: &topologyv1.LabelPolicy{
+				Columns:   []string{"display_name"},
+				Fallback:  "type_label",
+				MaxLength: 80,
+				Array:     "reject",
+			},
+			Modal: snmpTopologyV1EndpointModal(),
+		},
+	}
+	return types
+}
+
+func snmpTopologyV1DeviceModal() *topologyv1.ModalPresentation {
+	return &topologyv1.ModalPresentation{
+		Labels:       snmpTopologyV1DeviceModalLabels(),
+		MiniTopology: &topologyv1.ModalMiniTopologyPresentation{Depth: 1},
+		Sections: []topologyv1.ModalSection{
+			{
+				ID:    "ports",
+				Label: "Ports",
+				Order: 1,
+				Source: topologyv1.ModalSource{
+					Kind:  "actor_table",
+					Table: "actor_ports",
+				},
+				OwnerFilter: &topologyv1.ModalOwnerFilter{
+					Mode:        "actor_column",
+					ActorColumn: "actor",
+				},
+				Columns: snmpTopologyV1PortModalColumns(),
+				Sort:    &topologyv1.ModalSort{Column: "if_index", Direction: "asc"},
+			},
+			snmpTopologyV1PortLinksSection(2),
+		},
+	}
+}
+
+func snmpTopologyV1EndpointModal() *topologyv1.ModalPresentation {
+	return &topologyv1.ModalPresentation{
+		Labels:       snmpTopologyV1EndpointModalLabels(),
+		MiniTopology: &topologyv1.ModalMiniTopologyPresentation{Depth: 1},
+		Sections:     []topologyv1.ModalSection{snmpTopologyV1LinksSection(1)},
+	}
+}
+
+func snmpTopologyV1DeviceModalLabels() *topologyv1.ModalLabelsPresentation {
+	return &topologyv1.ModalLabelsPresentation{
+		Table: "actor_labels",
+		Identification: &topologyv1.ModalLabelIdentificationPresentation{
+			Fields: []topologyv1.ModalLabelIdentificationField{
+				{Key: "display_name", Label: "Name", MaxValues: 1},
+				{Key: "management_ip", Label: "Management IP", MaxValues: 1},
+				{Key: "vendor", Label: "Vendor", MaxValues: 1},
+				{Key: "model", Label: "Model", MaxValues: 1},
+				{Key: "ports_total", Label: "Ports", MaxValues: 1},
+				{Key: "lldp_neighbor_count", Label: "LLDP", MaxValues: 1},
+				{Key: "cdp_neighbor_count", Label: "CDP", MaxValues: 1},
+			},
+		},
+	}
+}
+
+func snmpTopologyV1EndpointModalLabels() *topologyv1.ModalLabelsPresentation {
+	return &topologyv1.ModalLabelsPresentation{
+		Table: "actor_labels",
+		Identification: &topologyv1.ModalLabelIdentificationPresentation{
+			Fields: []topologyv1.ModalLabelIdentificationField{
+				{Key: "display_name", Label: "Name", MaxValues: 1},
+				{Key: "ip_address", Label: "IP", MaxValues: 2},
+				{Key: "mac_address", Label: "MAC", MaxValues: 2},
+				{Key: "hostname", Label: "Hostname", MaxValues: 2},
+			},
+		},
+	}
+}
+
+func snmpTopologyV1LinksSection(order int) topologyv1.ModalSection {
+	return topologyv1.ModalSection{
+		ID:    "links",
+		Label: "Links",
+		Order: order,
+		Source: topologyv1.ModalSource{
+			Kind: "links",
+		},
+		OwnerFilter: &topologyv1.ModalOwnerFilter{
+			Mode:           "incident_link",
+			SrcActorColumn: "src_actor",
+			DstActorColumn: "dst_actor",
+		},
+		Columns: []topologyv1.ModalColumn{
+			{
+				ID:    "remote",
+				Label: "Remote Actor",
+				Projection: topologyv1.ModalProjection{
+					Kind:           "opposite_actor",
+					SrcActorColumn: "src_actor",
+					DstActorColumn: "dst_actor",
+				},
+				Cell: "actor_link",
+			},
+			modalSelectedSidePortColumn("local_port", "Local Port", "src_port_name", "dst_port_name"),
+			modalSelectedSidePortColumn("remote_port", "Remote Port", "dst_port_name", "src_port_name"),
+			modalDirectColumn("protocol", "Protocol", "protocol", "badge"),
+			modalDirectColumn("direction", "Direction", "direction", "text"),
+			modalDirectColumn("state", "State", "state", "badge"),
+			modalDirectColumn("evidence_count", "Evidence", "evidence_count", "number"),
+		},
+	}
+}
+
+func snmpTopologyV1PortLinksSection(order int) topologyv1.ModalSection {
+	return topologyv1.ModalSection{
+		ID:    "port_neighbors",
+		Label: "Port Neighbors",
+		Order: order,
+		Source: topologyv1.ModalSource{
+			Kind:  "actor_table",
+			Table: "actor_port_links",
+		},
+		OwnerFilter: &topologyv1.ModalOwnerFilter{
+			Mode:        "actor_column",
+			ActorColumn: "actor",
+		},
+		Columns:    snmpTopologyV1PortLinkModalColumns(),
+		Sort:       &topologyv1.ModalSort{Column: "if_index", Direction: "asc"},
+		EmptyLabel: "No port neighbors",
+	}
+}
+
+func modalSelectedSidePortColumn(id, label, selectedSrcPortColumn, selectedDstPortColumn string) topologyv1.ModalColumn {
+	return topologyv1.ModalColumn{
+		ID:    id,
+		Label: label,
+		Projection: topologyv1.ModalProjection{
+			Kind:             "selected_side_endpoint",
+			SrcActorColumn:   "src_actor",
+			DstActorColumn:   "dst_actor",
+			LocalPortColumn:  selectedSrcPortColumn,
+			RemotePortColumn: selectedDstPortColumn,
+		},
+		Cell: "text",
+	}
+}
+
+func modalDirectColumn(id, label, sourceColumn, cell string) topologyv1.ModalColumn {
+	return topologyv1.ModalColumn{
+		ID:         id,
+		Label:      label,
+		Projection: topologyv1.ModalProjection{Kind: "direct", Column: sourceColumn},
+		Cell:       cell,
+	}
+}
+
+func modalDirectColumnWithVisibility(id, label, sourceColumn, cell, visibility string) topologyv1.ModalColumn {
+	column := modalDirectColumn(id, label, sourceColumn, cell)
+	column.Visibility = visibility
+	return column
+}
+
+func modalActorRefColumn(id, label, actorColumn string) topologyv1.ModalColumn {
+	return topologyv1.ModalColumn{
+		ID:    id,
+		Label: label,
+		Projection: topologyv1.ModalProjection{
+			Kind:        "actor_ref_label",
+			ActorColumn: actorColumn,
+		},
+		Cell: "actor_link",
+	}
+}
+
+func modalActorRefColumnWithVisibility(id, label, actorColumn, visibility string) topologyv1.ModalColumn {
+	column := modalActorRefColumn(id, label, actorColumn)
+	column.Visibility = visibility
+	return column
+}
+
+func snmpTopologyV1PortTypes() map[string]topologyv1.PortType {
+	return map[string]topologyv1.PortType{
+		"lldp":           {Presentation: &topologyv1.PortPresentation{Label: "lldp/cdp", ColorSlot: "accent", Opacity: "normal"}},
+		"switch_facing":  {Presentation: &topologyv1.PortPresentation{Label: "switch-facing", ColorSlot: "primary", Opacity: "normal"}},
+		"host_facing":    {Presentation: &topologyv1.PortPresentation{Label: "host-facing", ColorSlot: "secondary", Opacity: "normal"}},
+		"host_candidate": {Presentation: &topologyv1.PortPresentation{Label: "host-candidate", ColorSlot: "info", Opacity: "normal"}},
+		"trunk":          {Presentation: &topologyv1.PortPresentation{Label: "trunk", ColorSlot: "warning", Opacity: "normal"}},
+		"access":         {Presentation: &topologyv1.PortPresentation{Label: "access", ColorSlot: "derived", Opacity: "normal"}},
+		"topology":       {Presentation: &topologyv1.PortPresentation{Label: "unclassified", ColorSlot: "neutral", Opacity: "normal"}},
+		"idle":           {Presentation: &topologyv1.PortPresentation{Label: "idle", ColorSlot: "muted", Opacity: "muted"}},
+		"unknown":        {Presentation: &topologyv1.PortPresentation{Label: "unknown", ColorSlot: "dim", Opacity: "muted"}},
+	}
+}
+
+type snmpTopologyV1LinkTypeSpec struct {
+	id           string
+	label        string
+	colorSlot    string
+	lineStyle    string
+	width        string
+	semanticRole string
+}
+
+func snmpTopologyV1LinkTypeSpecs() []snmpTopologyV1LinkTypeSpec {
+	return []snmpTopologyV1LinkTypeSpec{
+		{id: snmpTopologyV1LinkLLDP, label: "LLDP", colorSlot: "accent", lineStyle: "solid", width: "thick", semanticRole: "discovery"},
+		{id: snmpTopologyV1LinkCDP, label: "CDP", colorSlot: "accent", lineStyle: "solid", width: "thick", semanticRole: "discovery"},
+		{id: snmpTopologyV1LinkBridge, label: "Bridge", colorSlot: "neutral", lineStyle: "solid", width: "normal", semanticRole: "normal"},
+		{id: snmpTopologyV1LinkFDB, label: "FDB", colorSlot: "neutral", lineStyle: "solid", width: "normal", semanticRole: "normal"},
+		{id: snmpTopologyV1LinkSTP, label: "STP", colorSlot: "muted", lineStyle: "solid", width: "normal", semanticRole: "normal"},
+		{id: snmpTopologyV1LinkARP, label: "ARP", colorSlot: "muted", lineStyle: "solid", width: "normal", semanticRole: "normal"},
+		{id: snmpTopologyV1LinkSNMP, label: "SNMP", colorSlot: "primary", lineStyle: "solid", width: "normal", semanticRole: "normal"},
+		{id: snmpTopologyV1LinkProbable, label: "Probable", colorSlot: "dim", lineStyle: "solid", width: "normal", semanticRole: "normal"},
+		{id: snmpTopologyV1LinkObservation, label: "L2 observation", colorSlot: "neutral", lineStyle: "solid", width: "normal", semanticRole: "normal"},
+	}
+}
+
+func snmpTopologyV1LinkTypes() map[string]topologyv1.LinkType {
+	types := make(map[string]topologyv1.LinkType)
+	for _, spec := range snmpTopologyV1LinkTypeSpecs() {
+		types[spec.id] = topologyv1.LinkType{
+			Orientation:   "observed_bidirectional",
+			DirectionRole: "observation",
+			SemanticRole:  spec.semanticRole,
+			Aggregation: topologyv1.LinkAggregation{
+				Direction: "canonicalize_unordered",
+				Evidence:  "append",
+				Metrics: map[string]string{
+					"evidence_count": "sum",
+				},
+			},
+			EvidenceTypes: []string{spec.id},
+			Presentation: &topologyv1.LinkPresentation{
+				Label:     spec.label,
+				ColorSlot: spec.colorSlot,
+				LineStyle: spec.lineStyle,
+				Width:     spec.width,
+				Curve:     "straight",
+				Arrow:     "none",
+			},
+		}
+	}
+	return types
+}
+
+func snmpTopologyV1EvidenceTypes() map[string]topologyv1.EvidenceType {
+	types := make(map[string]topologyv1.EvidenceType)
+	for _, spec := range snmpTopologyV1LinkTypeSpecs() {
+		types[spec.id] = topologyv1.EvidenceType{
+			LinkType: spec.id,
+			Role:     "observation_evidence",
+			Columns:  snmpTopologyV1EvidenceColumns(),
+			MatchColumns: []string{
+				"src_actor",
+				"dst_actor",
+				"protocol",
+				"src_endpoint",
+				"dst_endpoint",
+			},
+		}
+	}
+	return types
+}
+
+func snmpTopologyV1Presentation() *topologyv1.Presentation {
+	return &topologyv1.Presentation{
+		ProfileVersion: "snmp-l2.v1",
+		Selection: &topologyv1.SelectionPresentation{
+			ActorClick: &topologyv1.ActorClickPresentation{Mode: "highlight_connections"},
+		},
+		Legend: &topologyv1.PresentationLegend{
+			Actors: []topologyv1.LegendEntry{
+				{Type: "router", Label: "Router"},
+				{Type: "switch", Label: "Switch"},
+				{Type: "firewall", Label: "Firewall"},
+				{Type: "access_point", Label: "Access Point"},
+				{Type: "server", Label: "Server"},
+				{Type: "storage", Label: "Storage"},
+				{Type: "load_balancer", Label: "Load Balancer"},
+				{Type: "printer", Label: "Printer"},
+				{Type: "phone", Label: "IP Phone"},
+				{Type: "ups", Label: "UPS / PDU"},
+				{Type: "camera", Label: "Camera / Media"},
+				{Type: "device", Label: "Other device"},
+				{Type: "custom", Label: "Other"},
+				{Type: "endpoint", Label: "Inferred endpoint"},
+				{Type: "segment", Label: "Network segment"},
+			},
+			Links: []topologyv1.LegendEntry{
+				{Type: snmpTopologyV1LinkLLDP, Label: "LLDP"},
+				{Type: snmpTopologyV1LinkCDP, Label: "CDP"},
+				{Type: snmpTopologyV1LinkSNMP, Label: "SNMP"},
+				{Type: snmpTopologyV1LinkBridge, Label: "Bridge"},
+				{Type: snmpTopologyV1LinkProbable, Label: "Probable"},
+			},
+			Ports: []topologyv1.LegendEntry{
+				{Type: "lldp", Label: "lldp/cdp"},
+				{Type: "switch_facing", Label: "switch-facing"},
+				{Type: "host_facing", Label: "host-facing"},
+				{Type: "host_candidate", Label: "host-candidate"},
+				{Type: "trunk", Label: "trunk"},
+				{Type: "access", Label: "access"},
+				{Type: "topology", Label: "unclassified"},
+				{Type: "idle", Label: "idle"},
+				{Type: "unknown", Label: "unknown"},
+			},
+		},
+		PortFields: []topologyv1.PresentationField{
+			{Key: "type", Label: "Type"},
+			{Key: "role", Label: "Role"},
+			{Key: "status", Label: "Status"},
+			{Key: "mode", Label: "Mode"},
+			{Key: "sources", Label: "Sources"},
+		},
+	}
+}
